+++ src/main/java/nextstep/app/WebConfig.java
@@ -1,14 +1,7 @@
 package nextstep.app;
 
-import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
-import nextstep.app.ui.BasicAuthenticationInterceptor;
-import nextstep.app.ui.FormLoginInterceptor;
-import nextstep.security.user.UserDetails;
-import nextstep.security.user.UserDetailsService;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
@@ -19,28 +12,28 @@ public class WebConfig implements WebMvcConfigurer {
         this.memberRepository = memberRepository;
     }
 
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new FormLoginInterceptor(userDetailsService())).addPathPatterns("/login");
-        registry.addInterceptor(new BasicAuthenticationInterceptor(userDetailsService())).addPathPatterns("/members");
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        return username -> {
-            Member member = memberRepository.findByEmail(username)
-                    .orElseThrow(() -> new IllegalArgumentException("해당하는 사용자를 찾을 수 없습니다."));
-            return new UserDetails() {
-                @Override
-                public String getUsername() {
-                    return member.getEmail();
-                }
-
-                @Override
-                public String getPassword() {
-                    return member.getPassword();
-                }
-            };
-        };
-    }
+//    @Override
+//    public void addInterceptors(InterceptorRegistry registry) {
+//        registry.addInterceptor(new FormLoginInterceptor(userDetailsService())).addPathPatterns("/login");
+//        registry.addInterceptor(new BasicAuthenticationInterceptor(userDetailsService())).addPathPatterns("/members");
+//    }
+//
+//    @Bean
+//    public UserDetailsService userDetailsService() {
+//        return username -> {
+//            Member member = memberRepository.findByEmail(username)
+//                    .orElseThrow(() -> new IllegalArgumentException("해당하는 사용자를 찾을 수 없습니다."));
+//            return new UserDetails() {
+//                @Override
+//                public String getUsername() {
+//                    return member.getEmail();
+//                }
+//
+//                @Override
+//                public String getPassword() {
+//                    return member.getPassword();
+//                }
+//            };
+//        };
+//    }
 }
