+++ src/main/java/nextstep/security/filter/FormLoginFilter.java
@@ -15,6 +15,7 @@ import java.util.Map;
 
 public class FormLoginFilter extends OncePerRequestFilter {
     public static final String SPRING_SECURITY_CONTEXT_KEY = "SPRING_SECURITY_CONTEXT";
+    private static final String MATCH_URI = "/login";
 
     private final UserDetailsService userDetailsService;
 
@@ -24,18 +25,24 @@ public class FormLoginFilter extends OncePerRequestFilter {
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        Map<String, String[]> parameterMap = request.getParameterMap();
-        String username = parameterMap.get("username")[0];
-        String password = parameterMap.get("password")[0];
-
-        UserDetails userDetails = userDetailsService.loadUserByUsername(username);
-        if (!userDetails.getPassword().equals(password)) {
-            throw new AuthenticationException();
+        if (!request.getRequestURI().equals(MATCH_URI)) {
+            filterChain.doFilter(request, response);
+            return;
         }
+        try {
+            Map<String, String[]> parameterMap = request.getParameterMap();
+            String username = parameterMap.get("username")[0];
+            String password = parameterMap.get("password")[0];
 
-        HttpSession session = request.getSession();
-        session.setAttribute(SPRING_SECURITY_CONTEXT_KEY, userDetails);
+            UserDetails userDetails = userDetailsService.loadUserByUsername(username);
+            if (!userDetails.getPassword().equals(password)) {
+                throw new AuthenticationException();
+            }
 
-        filterChain.doFilter(request, response);
+            HttpSession session = request.getSession();
+            session.setAttribute(SPRING_SECURITY_CONTEXT_KEY, userDetails);
+        } catch (Exception e) {
+            response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        }
     }
 }
